converted interaction class and extended with different types of employees
