--- src/main/java/com/businessName/ticketApi/FunctionalInterfaceExample.java
@@ -1,6 +1,0 @@
-package com.businessName.ticketApi;
-
-@FunctionalInterface
-public interface FunctionalInterfaceExample {
-    int math(int numOne);
-}
