+++ src/main/java/com/businessName/ticketService/EmployeeInteractions.java
@@ -1,6 +1,42 @@
 package com.businessName.ticketService;
 
-public interface EmployeeInteractions {
-    String doLogin(String jsonFromApi);
-    String updatePersonalInfo(String jsonFromApi);
+import com.businessName.dataEntity.DatabaseEntity;
+import com.businessName.ticketDao.DataAccessImp;
+import com.businessName.ticketDao.DataAccessInterface;
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+import java.util.HashMap;
+import java.util.Objects;
+
+public class EmployeeInteractions {
+
+    public DataAccessImp daoObject;
+
+    public EmployeeInteractions(DataAccessInterface daoObject) {
+        this.daoObject = (DataAccessImp) daoObject;
+    }
+
+    public String doLogin(String jsonFromApi) {
+        HashMap<String, String> loginMap = new Gson().fromJson(
+                String.valueOf(jsonFromApi),
+                new TypeToken<HashMap<String, String>>() {}.getType());
+        DatabaseEntity loginTech = new DatabaseEntity(loginMap);
+        loginTech.sanitizeFromApi();
+        DatabaseEntity[] employeeInfo = daoObject.selectObjectsDb(loginTech.selectDoLogin());
+        if(employeeInfo.length<1) {
+            return "Incorrect Username!";
+        } else if(Objects.equals(employeeInfo[0].newRowObject.get("pass"), loginMap.get("pass"))) {
+            return employeeInfo[0].newRowObject.get("username")+
+                "_"+employeeInfo[0].newRowObject.get("type_id")+
+                "_"+employeeInfo[0].newRowObject.get("employees_id");
+        } else {
+            System.out.println(employeeInfo[0].newRowObject.get("pass"));
+            System.out.println(loginMap.get("pass"));
+            return "Incorrect Password!";
+            }
+        }
+
+        public String updatePersonalInfo(String jsonFromApi) {
+        return null;
+        }
 }
