+++ src/main/java/com/businessName/ticketService/TechnicianInteractions.java
@@ -7,38 +7,10 @@ import com.google.gson.reflect.TypeToken;
 import java.util.HashMap;
 import java.util.Objects;
 
-public class TechnicianInteractions implements EmployeeInteractions {
-    public DataAccessInterface daoObject;
+public class TechnicianInteractions extends EmployeeInteractions {
 
     public TechnicianInteractions(DataAccessInterface daoObject) {
-        this.daoObject = daoObject;
-    }
-
-    @Override
-    public String doLogin(String jsonFromApi) {
-        HashMap<String, String> loginMap = new Gson().fromJson(
-                String.valueOf(jsonFromApi), new TypeToken<HashMap<String, String>>() {}.getType()
-        );
-        DatabaseEntity loginTech = new DatabaseEntity(loginMap);
-        loginTech.sanitizeFromApi();
-        DatabaseEntity[] employeeInfo = daoObject.selectObjectsDb(loginTech.selectDoLogin());
-        if(employeeInfo.length<1) {
-            return "Incorrect Username!";
-        } else if(Objects.equals(employeeInfo[0].newRowObject.get("pass"), loginMap.get("pass"))) {
-            return employeeInfo[0].newRowObject.get("username")+
-                    "_"+employeeInfo[0].newRowObject.get("type_id")+
-                    "_"+employeeInfo[0].newRowObject.get("employees_id");
-        } else {
-            System.out.println(employeeInfo[0].newRowObject.get("pass"));
-            System.out.println(loginMap.get("pass"));
-            return "Incorrect Password!";
-        }
-
-    }
-
-    @Override
-    public String updatePersonalInfo(String jsonFromApi) {
-        return null;
+        super(daoObject);
     }
 
     public String createHelpTicket(String jsonFromApi) { return null; }
